tests: arch: arm: move test arm_interrupt to new ztest API Migrate the testsuite tests/arch/arm/arm_interrupt to the new ztest API.
