+++ tests/arch/arm/arm_interrupt/prj.conf
@@ -3,3 +3,4 @@ CONFIG_DYNAMIC_INTERRUPTS=y
 CONFIG_TEST_USERSPACE=y
 CONFIG_APPLICATION_DEFINED_SYSCALL=y
 CONFIG_IDLE_STACK_SIZE=512
+CONFIG_ZTEST_NEW_API=y
